requirements.txt
fastapi==0.101.0
annotated-types==0.5.0
gunicorn==20.1.0
uvicorn==0.22.0
apscheduler==3.10.1
pymongo==4.4.1
python-dotenv==1.0.0
pyrebase4==4.7.1
firebase_admin==6.1.0
urllib3==1.26.16